port/requirements.txt
flask
pypdf
groq
python-dotenv
gunicorn
